+++ fltk-derive/src/window.rs
@@ -77,6 +77,8 @@ pub fn impl_window_trait(ast: &DeriveInput) -> TokenStream {
         format!("{}_{}", name_str, "default_cursor").as_str(),
         name.span(),
     );
+    let screen_num = Ident::new(format!("{}_{}", name_str, "screen_num").as_str(), name.span());
+    let set_screen_num = Ident::new(format!("{}_{}", name_str, "set_screen_num").as_str(), name.span());
 
     let gen = quote! {
         unsafe impl HasRawWindowHandle for #name {
@@ -384,6 +386,20 @@ pub fn impl_window_trait(ast: &DeriveInput) -> TokenStream {
                 unsafe {
                     #default_cursor(self.inner, cursor as i32)
                 }
+            }
+
+            fn screen_num(&self) -> i32 {
+                assert!(!self.was_deleted());
+                unsafe {
+                    #screen_num(self.inner)
+                }
+            }
+
+            fn set_screen_num(&mut self, n: i32) {
+                assert!(!self.was_deleted());
+                unsafe {
+                    #set_screen_num(self.inner, n)
+                } 
             }
         }
     };